# sort changes, faster