… tests/images/masks/strokeEllipse.png
… tests/images/svg/diffs/Ghostscript_Tiger.png
… tests/images/svg/diffs/ellipse01.png
… tests/images/svg/diffs/rect02.png
… tests/images/svg/rendered/Ghostscript_Tiger.png
… tests/images/svg/rendered/ellipse01.png
… tests/images/svg/rendered/rect02.png