before:
`svg decode ........................ 19.083 ms     19.282 ms    ±0.133   x259`

after:
`svg decode ........................ 18.788 ms     18.943 ms    ±0.120   x264`